fix: import issues with background workers
